Updated the rest of the read/write functions to use the exact way the clients read/write data.  As a result, maximum values for all of them (aside from short, which was already updated) have increased.  This may fix some problems, but I'd imagine they're rare if they exist.
Updated the server > client packet enumeration to include values that're valid packets in 5.07, but weren't listed.  I've also noted all packets that're listed, but aren't used in 5.07.

Added a UPNP preprocessor flag that allows universal plug and play to not be compiled with the server.  Also added another .bat file for Visual Studio 2010 for no UPNP.

Changed the "homepath" obtaining function to use GetCurrentDirectory (working directory) instead of using the executable directory.