Fix putnpc and join for GS1.
The function to minify NPC script works off of normal \n line breaks.  There were cases where the script was using \xa7 linebreaks, which would cause the minify function to break and erase the whole script.

This removes the conversion of \n to \xa7 in the `Server::addNpc` code since the minify function performs that duty.

This also fixes the `join` command, which was converting to \xa7 before being passed through minify.

Finally, this updates `Player::msgPLI_PUTNPC` to properly mark the NPC type as `PUTNPC` instead of `LEVELNPC`.