Replaced CLog with a new file logging system.
The new system replaces the printf() style of string placeholders with the modern std::format method.  It is extremely easy to add new file targets and the logging is very customizable.  Additionally, a std::formatter for CString is provided so CStrings can be printed as-is without needing to call CString::text().

Finally, all references to fmtlib have been replaced with std::format variants.