Fixing bugs caused by the refactors.
It turns out we actually DO need the know the base path of the executable; at least for V8, since it needs to know the location of some external data.  I've reintroduced it, but altered it to no longer pull from the current working directory, but rather where the executable is being ran from.

Additionally, fixes for NC and RC1 were implemented.  NC will log in properly and the file browser for RC1 displays directories.

And finally, some general bug fixes were included.