Re-enabled a lot of the npc-server security checks.
When the npc-server is active, a lot of actions from the client aren't trusted anymore and are dropped.  This reintroduces a lot of those security checks.

We also send PLO_HASNPCSERVER on client login, fully informing the client that it should switch to server-authoritative on a lot of its features.