Lots of scripting engine changes.
The SourceCode class no longer stores "server bytecode".  Instead, it stores the results of the script compilation from the script engine.  This gets passed to the engine when the source code is executed.  This lets it support non-bytecode engines.  Additionally, joined classes are stored in the compilation results and are not stored individually.

Script events and variable stores are implemented.  An event, like 'playerenters', is passed to an event queue along with the source object that generated it, and any additional arguments.  Variable stores provide a cross-script-engine way of manipulating variables.  If a script engine needs to save more state information, it should handle that itself.