Large scripting variable rewrite.
The new system is currently less performant, but it will allow proper integration with the server in the future.  Variables now remember their identifiers and can have getter/setter callbacks.  If a variable is grabbed directly from a data store, a setter callback is set so any writes get propagated back to the store.