Large rewrite of the scripting variables and containers.
Variable handling was redesigned and completely rewritten.  It is more performant, less buggy, and properly supports variable storage locations and object access.

Additionally, some math and string functions have been added and stubs for all known functions have been created.

The lexer has been improved to prevent ambiguity with strings and identifiers that only consist of a message code.  Blank identifier tokens are inserted before any command that requires an identifier so the parser can properly look up the variable.  The lexer also has the known variable store modifiers built-in.

The parser recognizes the `with()` statement now.  Some performance optimizations have also been added to speed up the speed at which it generates the AST.  The storage tokens are linked to the identifier so the correct storage location can be picked.