NPC-Server: Rewrote object properties.
Instead of object properties (like location, health, money, etc) being set in the variable store, entities in the SourceObjectType list can inherit from ISourceObject and define an interface to its properties.  If the variable finding routines can't find a variable in the object's store, it will see if a property exists.

Additionally, the GS1 source stack has been rewritten.  The original source and initiator are not added to the stack.  Instead, if the stack is exhausted, it will then try the original source / initiator, with an option to bias towards one or the other.  This fixes problems with message codes.

Finally, GameVariables will immediately update themselves in their constructor to prevent stale data bugs.

GS1: players[].attachid and players[].attachtype have been added.