Fixed NPCID generation and NPC prop issues.
The IdGenerator is now much better at what it does.  Also, NPCs are actually given the proper ID when created, which resolves problems with being unable to delete NPCs (regression).

Addditionally, the NPC props for the level and gmap x/y locations are not sent unless the NPC is actually on a gmap or the NPC has changed its level.