NPC-Server: Support the playerlogout event.
The NPC-Server now stores its own player list.  The GS1 engine now uses that player list when executing scripts.  When the player logs out, the server removes the player, but the NPC-Server holds onto the memory and cleans the player up after the logout event.  More functions will be moved to the NPC-Server in the future as the goal is to separate it from the server as much as possible.