Fixed issues with player sprites by using a common modTime for the server frame.
When a player's sprite gets manually changed, the client only sends over the change in gani and re-sends the old sprite, for some reason.

To resolve this, the current time is recorded at the start of the server frame and all prop changes will use this time.  This lets us check if the gani was changed this frame.  If it was, and it was set to a new "def[]" gani, and we also got a sprite change, we don't then overwrite the gani.