Modifications to scripts are propagated to players.
Changes to NPC scripts cause the NPC to be "deleted" and the props to be re-sent to the clients.  The clients have some sort of security that only allows a script to be sent once.

Changes to weapon scripts cause the weapon to be deleted and re-added to the player.

Changes to a class script cause both NPCs and weapons that join to that class to update their scripts.

Clientside code in classes is now properly sent to players.  There is no need to the clientside join hack (in fact, the hack has been disabled for npc-server enabled servers).