Rewrote Map handling.
Improved internal gmap handling to better mimic a single large level.  Serverside scripts will now see all the entities on the whole gmap instead of the ones only in the current level, and those entities will have global tile positions for the whole gmap.  Additionally, level NPCs are able to cross the level boundary when on a gmap.

Implemented the syncdistancex, syncdistancey, and syncbydistanceinside server options.