GS1: Implemented new commands.
Both `playerupgradestatus` and the `#G` message code have been added.

Added the `onwater2` and `tiletype` functions.

Added the `savelog2` and `setpm` commands.

The NPC-Server player will now respond to PMs.  When receiving PMs, it will trigger the "pm" event on the Control-NPC.

Fixed a regression with the `#e` message code that caused a length of -1 to be converted to 0.