Improved the NC warp command.
The NC warp command will no longer fail when you try to set the level name to a gmap.  It still isn't a good idea to do so (as there is no gmap x/y prop in the packet).

Additionally, the warp command will detect when the level is the same and just adjust the NPC's position props.  So if you leave the name as a gmap and just change the tile X/Y, the NPC will warp around in its current level.