NPC-Server: Improved sleeping state.
The sleeping state is toggled into after a full execution of the NPC scripts, which allows logout events to be processed.

NPC saving is also blocked while in a sleeping state.  The NPCs are saved when the last player exits the server and not constantly while the NPC-Server is sleeping.