NPC-Server: Expanded scripted level items to all item types.
If a class exists on the server that matches the name of a default item (except all gralats, which are combined under same "gralats" class), the server will spawn a new NPC for that item, just like it does with gralats.

Gralats, arrows, bombs, darts, and hearts can stack with existing NPCs, while all others spawn a unique NPC each time.

All items are now given the "ITEM" NPC type.