Fixed GS2 weapon packets.
When a weapon is registered on the client, PLO_LOADSCRIPT is sent with a header that includes the CRC32 hash of the bytecode.  The packet does not contain bytecode so the header length is not included.

PLO_NPCWEAPONSCRIPT is sent when the client requests the bytecode with PLI_UPDATESCRIPT.  The header does NOT contain the CRC32 hash, but the length IS included as a GSHORT.  The bytecode follows after the header.